killall.sh: use pidof instead of using sed on ps output to determine pids of dhcp clients as the regexp used could result in false positives. See #476524 for background.
r52825

+++ killall.sh
@@ -1,16 +1,16 @@
 #!/bin/sh
 # Killall for dhcp clients.
 
-# Use [] in sed /address/ to avoid matching sed command itself in ps output
-pids=$(ps ax | sed -n '/[u]dhcpc\|[d]hclient\|[p]ump/s/^[ ]*\([0-9]*\).*/\1/p')
+for client in dhclient udhcpc pump; do
+	pid=$(pidof $client) || true
+	[ "$pid" ] || continue
 
-for pid in $pids; do
-  if kill -0 $pid 2>/dev/null; then
-    kill -TERM $pid
-    sleep 1
-    # Still alive? Die!
-    if kill -0 $pid 2>/dev/null; then
-      kill -KILL $pid
-    fi
-  fi
+	if kill -0 $pid 2>/dev/null; then
+		kill -TERM $pid
+		sleep 1
+		# Still alive? Die!
+		if kill -0 $pid 2>/dev/null; then
+			kill -KILL $pid
+		fi
+	fi
 done
